Deleted instrumented actor and small refactor of the StatusHelper
