+++ serving-api/src/main/scala/com/stratio/sparkta/serving/api/actor/ClusterSparkStreamingContextActor.scala
@@ -18,7 +18,8 @@ package com.stratio.sparkta.serving.api.actor
 
 import java.io.File
 
-import akka.actor.ActorRef
+import akka.actor.{Actor, ActorRef}
+import akka.event.slf4j.SLF4JLogging
 import akka.pattern.ask
 import akka.util.Timeout
 import com.stratio.sparkta.driver.service.StreamingContextService
@@ -35,7 +36,7 @@ import org.apache.commons.lang.StringEscapeUtils
 import scala.collection.JavaConversions._
 import scala.concurrent.duration._
 import scala.sys.process._
-import scala.util.{Success, Failure, Try}
+import scala.util.{Failure, Success, Try}
 
 class ClusterSparkStreamingContextActor(policy: AggregationPoliciesModel,
                                         streamingContextService: StreamingContextService,
@@ -43,7 +44,9 @@ class ClusterSparkStreamingContextActor(policy: AggregationPoliciesModel,
                                         hdfsConfig: Config,
                                         zookeeperConfig: Config,
                                         detailConfig: Option[Config],
-                                        policyStatusActor: ActorRef) extends InstrumentedActor with SparktaSerializer {
+                                        policyStatusActor: ActorRef) extends Actor
+with SLF4JLogging
+with SparktaSerializer {
 
   implicit val timeout: Timeout = Timeout(3.seconds)
